% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CATCOLDQRAIN.R
\docType{data}
\name{CATCOLDQRAIN}
\alias{CATCOLDQRAIN}
\title{environmental data describing rainfall during the coldest quarter for Melanotaenia fluviatilis sampling sites in the Murray-Darling Basin}
\format{A text file with one value per population}
\usage{
CATCOLDQRAIN
}
\description{
environmental data describing rainfall during the coldest quarter for Melanotaenia fluviatilis sampling sites in the Murray-Darling Basin
}
\examples{
data(CATCOLDQRAIN)

}
\keyword{datasets}
