% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vcf_INFO.R
\name{clean_vcf_INFO}
\alias{clean_vcf_INFO}
\title{Parse and clean-up INFO Fields in a VCF File}
\usage{
clean_vcf_INFO(vcffile)
}
\arguments{
\item{vcffile}{A character string specifying the path to the input VCF file.}
}
\value{
Writes a cleaned VCF file to the specified or default path.
}
\description{
This function reads a VCF file, processes the INFO fields to remove any fields that are not consistently present 
across all records, and outputs a cleaned VCF file with "_clean" appended to the input file name before the suffix.
}
\examples{
# Assume you have a VCF file "my_input.vcf" in your working directory
clean_vcf_INFO("my_input.vcf")

}
