% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DelphinusXY.R
\name{DelphinusXY}
\alias{DelphinusXY}
\title{Sample locations of common dolphins in Australia and New Zealand}
\format{
A data frame with 21 rows and 3 columns:
\describe{
  \item{ID}{Site name}
  \item{X}{Longitude}
  \item{Y}{Latitude}
}
}
\description{
Sample locations of common dolphins in Australia and New Zealand
}
\examples{
data(DelphinusXY)
head(yourData)

}
