% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_tri.R
\name{extract_tri}
\alias{extract_tri}
\title{extract upper or lower triangle from a square matrix}
\usage{
extract_tri(square.matrix, result.file, upper = FALSE, keep.diag = TRUE)
}
\arguments{
\item{square.matrix}{a pairwise matrix object}

\item{result.file}{a character string name of your result file}

\item{upper}{a logical indicating that you want to keep the upper triangle (default=FALSE)}

\item{keep.diag}{a logical indicating that you want to keep the diagonal in your trimat (default=TRUE)}
}
\value{
a trimatrix object and .csv file
}
\description{
extract upper or lower triangle from a square matrix
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 # load example matrix
 example.mat <- matrix(c(0,3,5,6, 3,0,6,6, 5,6,0,3, 6,6,3,0), 
                nrow = 4, ncol = 4, dimnames = list(c("A", "B", "C", "D"), 
                c("A", "B", "C", "D")))

 # extract lower trimarix
 Lmat <- extract_tri(example.mat, "lowermat")
 
 # extract upper trimarix without the diagonal
 Umat <- extract_tri(example.mat, "uppermat", upper = TRUE, keep.diag = FALSE)
}
\author{
Chris Brauer
}
