% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWE_filter.R
\name{HWE_filter}
\alias{HWE_filter}
\title{A wrapper function for the hw.test function from pegas to filter a genind object for per locus Hardy-Weinberg equilibrium by population (i.e. remove loci out of HWE in >x% of populations at a specified FDR).}
\usage{
HWE_filter(GenInd.obj, perm, FDR_cut, pop_thresh, no_cores)
}
\arguments{
\item{GenInd.obj}{a genind object}

\item{perm}{the number of permutations for the monte carlo simulations}

\item{FDR_cut}{he false discovery rate threshold}

\item{pop_thresh}{the number (%) of populations each locus needs to be in HWE to be retained}

\item{no_cores}{the number of cores to use for the analysis}
}
\value{
FDRadjP.csv a table of FDR adjusted P values for each locus per population

OutOfHWE_SNPs.csv a table of SNPs out of HWE at the specified FDR in >pop_thresh populations

HWE_SNPs.csv a table of SNPs in HWE at the specified FDR in >pop_thresh populations

HWE_list a vector of SNPs in HWE suitable for input to the subset_snps function
}
\description{
A wrapper function for the hw.test function from pegas to filter a genind object for per locus Hardy-Weinberg equilibrium by population (i.e. remove loci out of HWE in >x% of populations at a specified FDR).
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 # load data
 data(rainbow.genind)

 #subset rainbow.genind to 100 loci
 gen100 <- rainbow.genind[ ,1:200]

 # run HWE_filter
 run1 <- HWE_filter(gen100, perm=999, FDR_cut=0.1, pop_thresh=50, no_cores=6)

}
\keyword{HWE}
