% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_genotypes.R
\name{impute_genotypes}
\alias{impute_genotypes}
\title{impute missing genotypes based on pop structure characterised using snmf}
\usage{
impute_genotypes(input.genotypes, K = NULL, no_cores = 1)
}
\arguments{
\item{input.genotypes}{A genind or genlight object}

\item{K}{the number of predetermined clusters in your data (suggest using smnf, ADMIXTURE or similar to explore first)}

\item{no_cores}{the number of cores to use for the analysis}
}
\value{
a genind or genlight object with missing data imputed
}
\description{
impute missing genotypes based on pop structure characterised using snmf
}
\details{
The function imputes missing genotypes based on ancestral population allele frequencies estimated using LEA::snmf()
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 # load example coordinate file
 data(rainbow.genind)

 # run analysis for K=6
 imputed.gi <- impute_genotypes(rainbow.genind, K = 6)
 
}
\author{
Chris Brauer
}
