% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDB_cols.R
\docType{data}
\name{MDB_cols}
\alias{MDB_cols}
\title{Custom color palette based on the Murray-Darling Basin landscape}
\format{
A named vector of RGB colors.
\describe{
  \item{Mudflat}{RGB(190/255, 175/255, 152/255)}
  \item{Silt}{RGB(144/255, 128/255, 101/255)}
  \item{Murray}{RGB(168/255, 160/255, 100/255)}
  \item{Mallee}{RGB(102/255, 101/255, 50/255)}
  \item{Eucalyptus}{RGB(94/255, 112/255, 60/255)}
  \item{Grassland}{RGB(200/255, 179/255, 59/255)}
  \item{Ochre}{RGB(200/255, 112/255, 0/255)}
  \item{Sunset}{RGB(200/255, 80/255, 37/255)}
  \item{Ironbark}{RGB(167/255, 47/255, 26/255)}
  \item{PinkGum}{RGB(222/255, 116/255, 116/255)}
  \item{MorningSky}{RGB(199/255, 225/255, 255/255)}
  \item{EveningSky}{RGB(149/255, 217/255, 255/255)}
  \item{SummerSky}{RGB(76/255, 149/255, 255/255)}
}
}
\source{
Custom defined for the melfuR package
}
\usage{
MDB_cols
}
\description{
Custom color palette based on the natural elements and features
of the Murray-Darling Basin landscape.
}
\keyword{datasets}
