% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mf_mdsXY.R
\docType{data}
\name{Mf_mdsXY}
\alias{Mf_mdsXY}
\title{Spatial coordinates for Melanotaenia fluviatilis sampling sites in the Murray-Darling Basin}
\format{A csv file with x and y coordinates in decimal degrees}
\usage{
Mf_mdsXY
}
\description{
Spatial coordinates for Melanotaenia fluviatilis sampling sites in the Murray-Darling Basin
}
\examples{
data(Mf_mdsXY)

}
\keyword{datasets}
