% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_data.R
\name{missing_data}
\alias{missing_data}
\title{This function calculates the percent of missing genotypes for each population from a genind object}
\usage{
missing_data(genind)
}
\arguments{
\item{input.file}{A genind object}
}
\description{
This function calculates the percent of missing genotypes for each population from a genind object
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 ## load the example data
 data(rainbow.genind)

 # calculate the percent polymorphic loci per population
 res <- missing_data(rainbow.genind)

}
\keyword{data}
\keyword{missing}
