% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_poly.R
\name{percent_poly}
\alias{percent_poly}
\title{This function calculates the percent polymorphic loci per population from a genind object}
\usage{
percent_poly(genind)
}
\arguments{
\item{input.file}{A genind object}
}
\description{
This function calculates the percent polymorphic loci per population from a genind object
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 ## load the example data
 data(rainbow.genind)

 # calculate the percent polymorphic loci per population
 res <- percent_poly(rainbow.genind)

}
\keyword{diversity}
\keyword{genetic}
\keyword{loci}
\keyword{polymorphic}
