% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polarise_ancestral_vcf.R
\name{polarise_ancestral_vcf}
\alias{polarise_ancestral_vcf}
\title{Polarise a VCF File setting the (supplied) Ancestral Allele as REF}
\usage{
polarise_ancestral_vcf(
  input_vcf,
  output_vcf = "output.vcf.gz",
  compress = NULL
)
}
\arguments{
\item{input_vcf}{Path to the input VCF file.}

\item{output_vcf}{Path to the output VCF file. If not provided, defaults to "output.vcf.gz".}

\item{compress}{Optional boolean indicating whether to compress the output file. If not provided, it will automatically detect based on the input file.}
}
\value{
The function writes the modified VCF file to the specified output path and returns 
a modified vcfR object to the environment
}
\description{
This function reads a VCF file, reorders the REF and ALT alleles based on the ancestral allele 
specified in the INFO field (ID=AA) (the AA INFO field must be present in the vcf), adjusts the REF, ALT and genotype (GT) fields accordingly, and writes 
the modified VCF to a new file.
}
\examples{
polarise_ancestral_vcf("input.vcf")

}
