% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow.env.R
\docType{data}
\name{rainbow.env}
\alias{rainbow.env}
\title{Environmental data for Melanotaenia fluviatilis sampling sites in the Murray-Darling Basin}
\format{A table with location names, x and y coordinates in decimal degrees and six PCA axes describing hydroclimatic variation}
\usage{
rainbow.env
}
\description{
Environmental data for Melanotaenia fluviatilis sampling sites in the Murray-Darling Basin
}
\examples{
data(rainbow.env)

}
\keyword{datasets}
