% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainbow.genind.R
\docType{data}
\name{rainbow.genind}
\alias{rainbow.genind}
\title{ddRAD data from Melanotaenia fluviatilis in the Murray-Darling Basin}
\format{
A genind object with 500 loci, 249 individuals from 14 populations
}
\usage{
rainbow.genind
}
\description{
This is a subset of 500 SNPs from the M. fluviatilis riverscape genomics study
}
\examples{
data(rainbow.genind)
pop <- rainbow.genind$pop
pop

}
\keyword{datasets}
