% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_angsd_results.R
\name{read_angsd_results}
\alias{read_angsd_results}
\title{Read ANGSD D-statistic results}
\usage{
read_angsd_results(filepath)
}
\arguments{
\item{filepath}{Path to the ANGSD result file (e.g., "ind_eachamD.res.txt").}
}
\value{
A data frame with parsed results and any all-NA columns removed.
}
\description{
This function imports a tab-delimited ANGSD D-statistics result file.
}
\details{
Rscript jackKnife.R file= result.abbababa indNames= inds.txt outfile= example_ANGSD_Dstat_res.txt
For some reason the result file often contains a column at the end that contains only NA values.
}
\examples{
# Load example file from package extdata
example_file <- system.file("extdata", "example_ANGSD_Dstat_res.txt", package = "melfuR")
res <- read_angsd_results(example_file)
head(res)

}
