% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.gINLAnd.R
\name{run.gINLAnd}
\alias{run.gINLAnd}
\title{Run gINLAnd genotype-environment association (GEA) analysis}
\usage{
run.gINLAnd(input.dir, res.file, cutoff, coord, env.var)
}
\arguments{
\item{x}{A Genepop file}
}
\description{
This function tests for GEAs using SNP data in a Genepop format
}
\examples{
 data(CATCOLDQRAIN)
 data(Mf_mdsXY)
 WD <- system.file("extdata", "rainbow.gen", package = "melfuR")
 run.gINLAnd("WD", "res.CATCOLDQRAIN", 10, Mf_mdsXY, CATCOLDQRAIN)
}
\keyword{gINLAnd}
