% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dapc.R
\name{test_dapc}
\alias{test_dapc}
\title{Run test DAPC on rainbow.genind}
\usage{
test_dapc(x)
}
\arguments{
\item{x}{A genind object}
}
\description{
This function tests the package by running a simple DAPC on the example data
}
\examples{
data(rainbow.genind)
test_dapc(rainbow.genind)
}
\keyword{dapc}
