% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_colors.R
\name{view_colors}
\alias{view_colors}
\title{View Color Palette}
\usage{
view_colors(colors)
}
\arguments{
\item{colors}{A named vector of colors.}
}
\description{
This function plots the MDB color palette.
}
\examples{
view_colors(MDB_cols)
}
