% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_list_to_excel.R
\name{write_list_to_excel}
\alias{write_list_to_excel}
\title{Write a List of Data Frames to an Excel File}
\usage{
write_list_to_excel(data_list, file_name)
}
\arguments{
\item{data_list}{A named list of data frames to be written to the Excel file. Each list element will be written to a separate sheet.}

\item{file_name}{A character string specifying the name of the output Excel file.}
}
\value{
This function does not return a value. It writes the list of data frames to an Excel file.
}
\description{
This function writes a list of data frames to an Excel file, with each element of the list being written to a separate sheet.
}
\examples{
\dontrun{
library(openxlsx)

# Create example data
data1 <- data.frame(A = 1:5, B = letters[1:5])
data2 <- data.frame(X = rnorm(5), Y = runif(5))

# Create a list of data frames
data_list <- list(Sheet1 = data1, Sheet2 = data2)

# Write the list to an Excel file
write_list_to_excel(data_list, "example.xlsx")
}
}
