% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy2kml.R
\name{xy2kml}
\alias{xy2kml}
\title{generate a kml file from dataframe of coordinates}
\usage{
xy2kml(sampleXY)
}
\arguments{
\item{sampleXY}{-  a data frame with three columns: ID, X and Y (in decimal degrees)}
}
\value{
a .kml file of that you can view in google earth
}
\description{
generate a kml file from dataframe of coordinates
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 # load example coordinate file
 data(Delphinus_XY)

 # run analysis
 xy2kml(Delphinus_XY)

}
\author{
Chris Brauer
}
