% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_stats.R
\name{diversity_stats}
\alias{diversity_stats}
\title{This function calculates basic genetic diversity stats from a genind object}
\usage{
diversity_stats(genind)
}
\arguments{
\item{input.file}{A genind object}
}
\description{
This function calculates basic genetic diversity stats from a genind object
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 ## load the example data
 data(rainbow.genind)

 # calculate Na, \%poly, Ho, He, and Fis (with 95\%CIs) per population
 res <- diversity_stats(rainbow.genind)

}
