% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.genotypes.R
\name{impute.data}
\alias{impute.data}
\title{impute missing genotypes based on pop structure characterised using snmf}
\usage{
impute.data(input.file, K = NULL, no_cores = 1)
}
\arguments{
\item{input.file}{A genind object}

\item{K}{the number of predetermined clusters in your data (suggest using smnf, ADMIXTURE or similar to explore first)}

\item{no_cores}{the number of cores to use for the analysis}
}
\value{
a genind object with missing data imputed
}
\description{
impute missing genotypes based on pop structure characterised using snmf
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 # load example coordinate file
 data(rainbow.genind)

 # run analysis for K=6
 imputed.gi <- impute.data(rainbow.genind, K = 6)
 
}
\author{
Chris Brauer
}
