% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.gINLAnd.R
\name{run.gINLAnd}
\alias{run.gINLAnd}
\title{gINLAnd genotype-environment association (GEA) analysis}
\usage{
run.gINLAnd(input.file, result.file, cutoff, coord, env.var)
}
\arguments{
\item{input.file}{A genind object}

\item{result.file}{name of your result file}

\item{cutoff}{logBF threshold for candidate loci (i.e. loci with logBF>cutoff will be considered candidates)}

\item{coord}{dataframe with x and y spatial coordinates for each population}

\item{env.var}{value of some environmental variable for each population}
}
\description{
A wrapper function to run gINLAnd using SNP data in ADEGENET's genind format
}
\examples{
 ## set directory for results to be written
 setwd("path/to/working/directory")

 ## load the example data
 data(rainbow.genind)
 data(rainbow.env)

 # get xy coordinates
 coords <- rainbow.env[,2:3]

 # get environmental variable
 tempPC1 <- as.data.frame(rainbow.env[,4])

 ## run gINLAnd
 run.gINLAnd(rainbow.genind, "res.tempPC1", 10, coords, tempPC1)
}
\references{
Guillot G, Vitalis R, le Rouzic A, Gautier M (2014) Detecting correlation between allele frequencies
                 and environmental variables as a signature of selection. A fast computational approach for
                 genome-wide studies. Spatial Statistics 8, 145–155.
}
\keyword{gINLAnd}
